+++ 3D/Prudent3.cpp
@@ -9,13 +9,13 @@ using namespace std;
 class Walker {
 public:
     Walker() : pos(3, 0), Max(3, INT_MIN), Min(3, INT_MAX) {
-        for (int i = 0; i < 3; ++i) {
+        for (unsigned i = 0; i < 3; ++i) {
             Max[i].put({0, 0}, 0);
             Min[i].put({0, 0}, 0);
         }
     }
 
-    int step(int dir, int add) {
+    int step(unsigned dir, int add) {
         if (add == 0) { return 0; }
         if (add > 0) {
             if (pos[dir] < Max[dir].at({pos[(dir + 1) % 3], pos[(dir + 2) % 3]})) return 0;
@@ -30,7 +30,7 @@ public:
     }
 
     void reframe() {
-        for (int i = 0; i < 3; ++i) {
+        for (unsigned i = 0; i < 3; ++i) {
             if (pos[i] > Max[i].at({pos[(i + 1) % 3], pos[(i + 2) % 3]})) Max[i].put({pos[(i + 1) % 3], pos[(i + 2) % 3]}, pos[i]);
             if (pos[i] < Min[i].at({pos[(i + 1) % 3], pos[(i + 2) % 3]})) Min[i].put({pos[(i + 1) % 3], pos[(i + 2) % 3]}, pos[i]);
         }
@@ -38,7 +38,7 @@ public:
 
     int norm1() const { return abs(pos[0]) + abs(pos[1]) + abs(pos[2]); }
 
-    int                    length = 0;
+    size_t                 length = 0;
     vector<int>            pos;
     vector<TriMatrix<int>> Max, Min;
 };
@@ -50,13 +50,13 @@ ostream &operator<<(ostream &o, const Walker &W) {
 
 int main(int argc, char **argv) {
     Hub         H("3D prudent walker", argc, argv, "l=1000,n=1");
-    int         l = H['l'], n = H['n'];
+    size_t      l = H['l'], n = H['n'];
     vector<int> ends(l, 0);
 
     {
         ProgressBar PB(n * l);
-        int64_t     t = 0;
-        for (int i = 0; i < n; ++i) {
+        size_t      t = 0;
+        for (size_t i = 0; i < n; ++i) {
             Walker W;
             while (W.length < l) {
                 if (n == 1) cout << W;
@@ -68,6 +68,6 @@ int main(int argc, char **argv) {
     }
 
     if (n > 1) {
-        for (int i = 0; i < l; ++i) cout << i << ' ' << double(ends[i]) / double(n) << endl;
+        for (size_t i = 0; i < l; ++i) cout << i << ' ' << double(ends[i]) / double(n) << endl;
     }
 }
