+++ tests/test_permutation.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 int main(int argc, char **argv) {
     Hub H("Streams and permutations", argc, argv, "n=10");
-    for (int k = 1; k <= int(H['n']); ++k) {
+    for (size_t k = 1; k <= size_t(H['n']); ++k) {
         spdlog::info("n: {:>2}  permutations: {:>7}  partitions: {:>2}", k, size(permutations(k)), size(partitions(k)));
     }
 }
